Refactor DatasetVisualization component and update imports in DatasetsContent and ModelsContent
