--- DashAI/front/src/components/models/DatasetVisualization.jsx
@@ -1,349 +1,0 @@
-import { useCallback, useState, useEffect } from "react";
-import {
-  Button,
-  Grid,
-  Typography,
-  CircularProgress,
-  Box,
-  Alert,
-  Divider,
-  Tabs,
-  Tab,
-} from "@mui/material";
-import { AddCircleOutline as AddIcon } from "@mui/icons-material";
-import {
-  getDatasetFile,
-  getDatasetInfo,
-  getDatasetFileFiltered,
-} from "../../api/datasets";
-import JobQueueWidget from "../jobs/JobQueueWidget";
-import { getDatasetStatus } from "../../utils/datasetStatus";
-import { formatDate } from "../../pages/results/constants/formatDate";
-import Header from "../notebooks/dataset/header/Header";
-import Tooltip from "@mui/material/Tooltip";
-import OverviewTab from "../notebooks/dataset/tabs/OverviewTab";
-import { NumericTab } from "../notebooks/dataset/tabs/NumericTab";
-import { CategoricalTab } from "../notebooks/dataset/tabs/CategoricalTab";
-import QualityTab from "../notebooks/dataset/tabs/QualityTab";
-import CorrelationsTab from "../notebooks/dataset/tabs/CorrelationsTab";
-import { QualityAlerts } from "../notebooks/dataset/QualityAlerts";
-
-export default function DatasetVisualization({
-  dataset,
-  onSessionCreated,
-  onNewSession,
-  existingSessions = [],
-  tasks = [],
-}) {
-  if (!dataset) {
-    return (
-      <Box
-        sx={{ display: "flex", justifyContent: "center", alignItems: "center" }}
-      >
-        <CircularProgress sx={{ color: "#00BEBB" }} />
-        <Typography>Loading...</Typography>
-      </Box>
-    );
-  }
-
-  const [datasetInfo, setDatasetInfo] = useState(null);
-  const [tab, setTab] = useState(0);
-
-  useEffect(() => {
-    setTab(0);
-    const fetchDatasetInfo = async () => {
-      if (isProcessing) return;
-
-      try {
-        const info = await getDatasetInfo(dataset.id);
-        setDatasetInfo(info);
-      } catch (error) {
-        setDatasetInfo(null);
-      }
-    };
-
-    fetchDatasetInfo();
-  }, [dataset.id, dataset.status]);
-
-  // fetchPage compatible with server-side filtering
-  const fetchDatasetPage = useCallback(
-    async (page, pageSize, filterModel) => {
-      if (isProcessing) return { rows: [], total: 0 };
-      try {
-        // Use getDatasetFile if no filters, else use getDatasetFileFiltered
-        const hasFilters =
-          filterModel &&
-          Array.isArray(filterModel.items) &&
-          filterModel.items.length > 0;
-        let data;
-        if (hasFilters) {
-          data = await getDatasetFileFiltered(
-            dataset.file_path,
-            page,
-            pageSize,
-            filterModel,
-          );
-        } else {
-          data = await getDatasetFile(dataset.file_path, page, pageSize);
-        }
-        return { rows: data.rows ?? [], total: data.total ?? 0 };
-      } catch (error) {
-        return { rows: [], total: 0 };
-      }
-    },
-    [dataset.file_path, dataset.status, dataset.id],
-  );
-
-  const status = getDatasetStatus(dataset.status);
-  const isProcessing = !(status === "Finished" || status === "Error");
-
-  return (
-    <>
-      <Box
-        sx={{ display: "flex", flexDirection: "column", gap: 1, mt: 2, mx: 1 }}
-      >
-        {/* Quick Stats Section */}
-        {!isProcessing && datasetInfo && (
-          <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
-            {/* Dataset quality score */}
-            <Box
-              sx={{
-                display: "flex",
-                justifyContent: "space-between",
-                alignItems: "center",
-                flexWrap: "wrap",
-              }}
-            >
-              <Box sx={{ display: "flex", flexDirection: "column", gap: 0.5 }}>
-                <Typography variant="h4">{dataset.name}</Typography>
-              </Box>
-              <Box>
-                <Tooltip
-                  title="Data Quality Score is calculated based on various factors including missing values, duplicate rows, and data consistency. A higher score indicates better data quality."
-                  arrow
-                >
-                  <Alert
-                    severity={
-                      datasetInfo?.quality_info?.data_quality_score >= 80
-                        ? "success"
-                        : datasetInfo?.quality_info?.data_quality_score >= 50
-                          ? "warning"
-                          : datasetInfo?.quality_info?.data_quality_score
-                            ? "error"
-                            : "info"
-                    }
-                    sx={{
-                      fontSize: "1rem",
-                      height: "45px",
-                      display: "flex",
-                      alignItems: "center",
-                      p: "0 12px",
-                    }}
-                  >
-                    Quality Score:{" "}
-                    {datasetInfo?.quality_info?.data_quality_score?.toFixed(
-                      2,
-                    ) ?? "N/A"}
-                    {datasetInfo?.quality_info?.data_quality_score ? "%" : ""}
-                  </Alert>
-                </Tooltip>
-              </Box>
-            </Box>
-            <Box
-              sx={{
-                display: "flex",
-                justifyContent: "space-between",
-                mb: 1,
-                flexWrap: "wrap",
-              }}
-            >
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <Typography variant="subtitle1" color="text.secondary">
-                  {formatDate(dataset.created)}
-                </Typography>
-              </Box>
-
-              {/* Buttons - Only New Session for Models module */}
-              <Box
-                sx={{
-                  display: "flex",
-                  alignItems: "flex-end",
-                  gap: 2,
-                  flexDirection: "column",
-                }}
-              >
-                <Grid
-                  sx={{
-                    minHeight: "40px",
-                    display: "flex",
-                    gap: 2,
-                    flexWrap: "wrap",
-                    justifyContent: "flex-start",
-                  }}
-                >
-                  <Button
-                    variant="contained"
-                    disabled={isProcessing || !tasks || tasks.length === 0}
-                    onClick={() => {
-                      if (onNewSession) {
-                        onNewSession();
-                      }
-                    }}
-                    endIcon={<AddIcon />}
-                    sx={{ height: "40px" }}
-                  >
-                    New Session
-                  </Button>
-                </Grid>
-              </Box>
-            </Box>
-            <Header
-              totalRows={datasetInfo?.total_rows}
-              totalColumns={datasetInfo?.total_columns}
-              fileSize={datasetInfo?.general_info?.memory_usage_mb}
-              duplicateRows={datasetInfo?.general_info?.duplicate_rows}
-              missingValues={datasetInfo?.nan}
-            />
-            {/* Data Quality Alerts */}
-            <Box>
-              <QualityAlerts
-                qualityInfo={datasetInfo?.quality_info}
-                generalInfo={datasetInfo?.general_info}
-                missingValues={datasetInfo?.nan}
-              />
-            </Box>
-            {/* Tabs */}
-            <Tabs
-              sx={{
-                bgcolor: "#2C2C2C",
-                borderRadius: 1,
-                minHeight: "48px",
-                "& .MuiTabs-indicator": {
-                  height: "2px",
-                },
-                "& .MuiTab-root": {
-                  minHeight: "48px",
-                  fontSize: "0.85rem",
-                  borderRadius: "4px",
-                  transition: "all 0.2s",
-                  border: "1px solid transparent",
-                  textTransform: "none",
-                  "&:hover": {
-                    bgcolor: "rgba(255,255,255,0.05)",
-                  },
-                  "&.Mui-disabled": {
-                    color: "rgb(150, 150, 150)",
-                    bgcolor: "rgb(32, 32, 32)",
-                    borderColor: "rgb(39, 39, 42)",
-                    opacity: 0.6,
-                    cursor: "not-allowed",
-                    filter: "grayscale(0.6)",
-                    position: "relative",
-                    "&::after": {
-                      content: '""',
-                      position: "absolute",
-                      inset: 0,
-                      borderRadius: "4px",
-                      pointerEvents: "none",
-                      background:
-                        "repeating-linear-gradient(45deg, transparent, transparent 10px, rgba(0,0,0,0.1) 10px, rgba(0,0,0,0.1) 20px)",
-                    },
-                  },
-                },
-              }}
-              value={tab}
-              onChange={(_, newValue) => setTab(newValue)}
-            >
-              <Tab label="Overview" />
-              <Tab
-                label="Numerical Analysis"
-                disabled={
-                  !datasetInfo?.numeric_stats ||
-                  Object.keys(datasetInfo.numeric_stats).length === 0
-                }
-              />
-              <Tab
-                label="Categorical"
-                disabled={
-                  !datasetInfo?.categorical_stats ||
-                  Object.keys(datasetInfo.categorical_stats).length === 0
-                }
-              />
-              <Tab label="Data Quality" disabled={!datasetInfo?.quality_info} />
-              <Tab
-                label="Correlations"
-                disabled={
-                  !datasetInfo?.correlations ||
-                  Object.keys(datasetInfo.correlations).length === 0
-                }
-              />
-            </Tabs>
-
-            {/* Divider */}
-            <Divider sx={{ my: 2 }} />
-
-            {/* Content based on selected tab */}
-            {tab === 0 && (
-              <OverviewTab
-                dataset={dataset}
-                dtypes={datasetInfo?.general_info?.dtypes}
-                nan={datasetInfo?.nan}
-                total_rows={datasetInfo?.total_rows}
-                fetchDatasetPage={fetchDatasetPage}
-              />
-            )}
-            {tab === 1 && (
-              <NumericTab numericStats={datasetInfo?.numeric_stats} />
-            )}
-            {tab === 2 && (
-              <CategoricalTab
-                categoricalStats={datasetInfo?.categorical_stats}
-              />
-            )}
-            {tab === 3 && (
-              <QualityTab
-                qualityInfo={datasetInfo?.quality_info}
-                totalRows={datasetInfo?.total_rows}
-              />
-            )}
-            {tab === 4 && (
-              <CorrelationsTab correlations={datasetInfo?.correlations} />
-            )}
-          </Box>
-        )}
-
-        {/* Processing placeholder */}
-        {isProcessing && (
-          <Box
-            sx={{
-              display: "flex",
-              justifyContent: "center",
-              alignItems: "center",
-              minHeight: 200,
-              flexDirection: "column",
-              gap: 2,
-            }}
-          >
-            <CircularProgress sx={{ color: "#00BEBB" }} />
-            <Typography>Processing your dataset...</Typography>
-            <Typography
-              variant="body2"
-              color="text.secondary"
-              textAlign="center"
-            >
-              This may take a few moments depending on the size of your data.
-            </Typography>
-          </Box>
-        )}
-      </Box>
-
-      <JobQueueWidget />
-    </>
-  );
-}
